Separate EMX importer from importer framework
