+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/emx/EmxImportJob.java
@@ -3,6 +3,7 @@ package org.molgenis.data.importer.emx;
 import org.molgenis.data.DatabaseAction;
 import org.molgenis.data.RepositoryCollection;
 import org.molgenis.data.importer.EntityImportReport;
+import org.molgenis.data.importer.MetaDataChanges;
 import org.molgenis.data.importer.ParsedMetaData;
 
 /**
